+++ test/test_sdc_agg_funcs.py
@@ -63,3 +63,18 @@ def test_get_statsmodel_dof_no_attribute():
 
     with pytest.raises(AttributeError, match="model does not have df_resid attribute"):
         get_statsmodel_dof(FakeModel())
+
+def test_agg_values_are_same_empty_series() -> None:
+    """Agg_values_are_same returns True for an empty Series (line 78)."""
+    from acro.sdc_agg_funcs import agg_values_are_same  # noqa: PLC0415
+
+    result = agg_values_are_same(pd.Series([], dtype=float))
+    assert result is True
+
+
+def test_agg_top_2_sum_non_numeric() -> None:
+    """Agg_top_2_sum returns 0 for a non-numeric Series (lines 118-119)."""
+    from acro.sdc_agg_funcs import agg_top_2_sum  # noqa: PLC0415
+
+    result = agg_top_2_sum(pd.Series(["a", "b", "c"]))
+    assert result == 0
